feat: adicionando validaçõespara produtos

+++ src/services/Product.service.ts
@@ -1,5 +1,6 @@
-import Products from '../database/models/products.model'
-import { IproductsDTO, Iproducts } from '../interfaces'
+import Products from '../database/models/products.model';
+import { IproductsDTO, Iproducts } from '../interfaces';
+import { NotFound, BadRequest } from '../errors';
 
 class ProductService {
   private Products = Products;
@@ -11,6 +12,9 @@ class ProductService {
 
   public async getProduct(id: number): Promise<Iproducts | null> {
     const product = await this.Products.findOne({ where: { id }});
+    if (!product) {
+      throw new NotFound('Product not found');
+    }
     return product;
   }
 
@@ -22,12 +26,23 @@ class ProductService {
 
   public async editProduct(id: number, dataProduct: IproductsDTO): Promise<Iproducts | null> {
     const { name, type, price, quantity } = dataProduct;
+
+    const product = await this.Products.findOne({ where: { id }});
+    if (!product) {
+      throw new NotFound('Product not found');
+    }
+    
     await this.Products.update({ name, type, price, quantity }, { where: { id } });
     const edited = await this.Products.findOne({ where: { id }});
     return edited;
   } 
 
   public async excludeProduct(id: number): Promise<boolean | null> {
+    const product = await this.Products.findOne({ where: { id }});
+    if (!product) {
+      throw new NotFound('Product not found');
+    }
+
     await this.Products.destroy({ where: { id } });
     return true;
   }
